Crate os

The main module and entrypoint

The operating system and app also starts in this module. Kernel code starts executing from entry.asm, after which rust_main() is called to initialize various pieces of functionality clear_bss(). (See its source code for details.)

We then call println! to display Hello, world!. The main module and entrypoint

The operating system and app also starts in this module. Kernel code starts executing from entry.asm, after which rust_main() is called to initialize various pieces of functionality clear_bss(). (See its source code for details.)

We then call println! to display Hello, world!.

Modules

console 🔒

SBI console driver, for text output

lang_items 🔒

The panic handler

logging 🔒

sbi 🔒

timer 🔒

Timer module for getting system time

Macros

print

print string macro (without timestamp, for internal/raw output)

println

println string macro with timestamp

Functions

clear_bss

clear BSS segment

rust_main

the rust entry-point of os