os

Module fs

File system in os

Modules

inode 🔒

Arc<Inode> -> OSInodeInner: In order to open files concurrently we need to wrap Inode into Arc,but Mutex in Inode prevents file systems from being accessed simultaneously

stdio 🔒

Stdin & Stdout

Structs

OSInode

A wrapper around a filesystem inode to implement File trait atop

OpenFlags

Open file flags

Stdin

Standard input

Stdout

Standard output

Traits

File

File trait

Functions

list_apps

List all files in the filesystems

open_file

Open file with flags