os

Module mm

Memory management implementation

x86-64 4-level paging, plus everything about memory management: frame allocator, page table, map area and memory set.

Re-exports

pub use address::PhysAddr;

pub use address::PhysPageNum;

pub use address::StepByOne;

pub use address::VirtAddr;

pub use address::VirtPageNum;

pub use address::phys_to_virt;

pub use frame_allocator::FrameTracker;

pub use frame_allocator::frame_alloc;

pub use frame_allocator::frame_dealloc;

pub use memory_set::KERNEL_SPACE;

pub use memory_set::MapPermission;

pub use memory_set::MemorySet;

pub use memory_set::ioremap;

pub use memory_set::kernel_token;

pub use page_table::PageTable;

pub use page_table::PageTableEntry;

pub use page_table::UserBuffer;

pub use page_table::UserBufferIterator;

pub use page_table::translated_byte_buffer;

pub use page_table::translated_ref;

pub use page_table::translated_refmut;

pub use page_table::translated_str;

Modules

address 🔒

Implementation of physical and virtual address and page number.

frame_allocator 🔒

Implementation of FrameAllocator which controls all the frames in the operating system.

heap_allocator 🔒

The global allocator

memory_set 🔒

Implementation of MapArea and MemorySet.

page_table 🔒

Implementation of PageTableEntry and PageTable.

Functions

init

initiate heap allocator, frame allocator and kernel space