os::syscall

Module thread

Thread management syscalls (thread_create/gettid/waittid).

Functions

sys_gettid

Get the tid of the current thread.

sys_thread_create

Create a new thread of the current process, with entry point entry and argument arg.

sys_waittid

Wait for the thread tid of the current process.