os::task

Module manager

Implementation of TaskManager and the pid -> process table.

Structs

PID2PROCESS

a pid -> process map, used to send signals and reap a specific process

TASK_MANAGER

TaskManager

A ready queue of TaskControlBlocks.

Functions

add_task

Interface offered to add a thread.

fetch_task

Interface offered to pop the first thread.

insert_into_pid2process

Insert a process into the pid -> process map.

pid2process

Look up a process by pid.

remove_from_pid2process

Remove a process from the pid -> process map (when it is reaped).

remove_task

Remove a thread from the ready queue (used when a process terminates).

wakeup_task

Wake up a blocked thread: mark it ready and put it back in the queue.