os::task::signal

Struct SignalFlags

pub struct SignalFlags(<SignalFlags as PublicFlags>::Internal);

Tuple Fields
0: <SignalFlags as PublicFlags>::Internal
Implementations

impl SignalFlags

pub const SIGDEF: Self

pub const SIGHUP: Self

pub const SIGINT: Self

pub const SIGQUIT: Self

pub const SIGILL: Self

pub const SIGTRAP: Self

pub const SIGABRT: Self

pub const SIGBUS: Self

pub const SIGFPE: Self

pub const SIGKILL: Self

pub const SIGUSR1: Self

pub const SIGSEGV: Self

pub const SIGUSR2: Self

pub const SIGPIPE: Self

pub const SIGALRM: Self

pub const SIGTERM: Self

pub const SIGSTKFLT: Self

pub const SIGCHLD: Self

pub const SIGCONT: Self

pub const SIGSTOP: Self

pub const SIGTSTP: Self

pub const SIGTTIN: Self

pub const SIGTTOU: Self

pub const SIGURG: Self

pub const SIGXCPU: Self

pub const SIGXFSZ: Self

pub const SIGVTALRM: Self

pub const SIGPROF: Self

pub const SIGWINCH: Self

pub const SIGIO: Self

pub const SIGPWR: Self

pub const SIGSYS: Self

impl SignalFlags

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u32

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesnโt correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove wonโt truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference wonโt truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl SignalFlags

pub const fn iter(&self) -> Iter<SignalFlags>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<SignalFlags>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl SignalFlags

pub fn check_error(&self) -> Option<(i32, &'static str)>