os::task::task

Struct TaskControlBlockInner

pub struct TaskControlBlockInner {

    pub trap_cx_ptr: usize,
    pub base_size: usize,
    pub task_cx: TaskContext,
    pub task_status: TaskStatus,
    pub memory_set: MemorySet,
    pub parent: Option<Weak<TaskControlBlock>>,
    pub children: Vec<Arc<TaskControlBlock>>,
    pub exit_code: i32,
    pub fd_table: Vec<Option<Arc<dyn File + Send + Sync>>>,
    pub signals: SignalFlags,
    pub signal_mask: SignalFlags,
    pub handling_sig: isize,
    pub signal_actions: SignalActions,
    pub killed: bool,
    pub frozen: bool,
    pub trap_ctx_backup: Option<TrapContext>,

}

Fields
trap_cx_ptr: usize

virtual address of the TrapContext on this process’s kernel stack. The kernel stack lives in the shared kernel high-half mapping, so this address is valid under any active page table.

base_size: usizetask_cx: TaskContexttask_status: TaskStatusmemory_set: MemorySetparent: Option<Weak<TaskControlBlock>>children: Vec<Arc<TaskControlBlock>>exit_code: i32fd_table: Vec<Option<Arc<dyn File + Send + Sync>>>signals: SignalFlags

signals received by the current process (pending)

signal_mask: SignalFlags

signals that should not be responded to

handling_sig: isize

the signal currently being handled by the user handler, -1 if none

signal_actions: SignalActions

the actions of the current process for all signals

killed: bool

true if the task is killed by a signal (or SIGKILL)

frozen: bool

true if the task is frozen by SIGSTOP, waiting for SIGCONT

trap_ctx_backup: Option<TrapContext>

the trap context saved when a user signal handler is invoked, restored by sigreturn

Implementations

impl TaskControlBlockInner

pub fn get_trap_cx(&self) -> &'static mut TrapContext

pub fn get_user_token(&self) -> usize

fn get_status(&self) -> TaskStatus

pub fn is_zombie(&self) -> bool

pub fn alloc_fd(&mut self) -> usize