Crate os

The main module and entrypoint

Various facilities of the kernels are implemented as submodules. The most important ones are:

[trap]: Handles all cases of switching from userspace to the kernel

[syscall]: System call handling and implementation

The operating system also starts in this module. Kernel code starts executing from entry.asm, after which [rust_main()] is called to initialize various pieces of functionality. (See its source code for details.)

We then call batch::run_next_app() and for the first time go to userspace.

Re-exports

pub use heap_allocator::init_heap;

pub use crate::frame_allocater::*;

Modules

batch

batch subsystem

config

console ğŸ”’

frame_allocater

heap_allocator

lang_items ğŸ”’

logging ğŸ”’

sync ğŸ”’

Synchronization and interior mutability primitives

syscall

Implementation of syscalls

Macros

print

æ‰“å°ç»™å®šè¡¨è¾¾å¼çš„å€¼åˆ°æ ‡å‡†è¾“å‡ºã€‚

println

æ‰“å°ç»™å®šè¡¨è¾¾å¼çš„å€¼åˆ°æ ‡å‡†è¾“å‡ºå¹¶æ¢è¡Œã€‚

Structs

PageAllocImpl

Statics

HEAP_ALLOCATOR ğŸ”’

Functions

kernel_interrupt ğŸ”’

kernel interrupt

main ğŸ”’

the rust entry-point of os