os::mm

Module frame_allocator

Implementation of FrameAllocator which controls all the frames in the operating system.

Structs

FRAME_ALLOCATOR

frame allocator instance through lazy_static!

FrameTracker

manage a frame which has the same lifecycle as the tracker

StackFrameAllocator

an implementation for frame allocator

Traits

FrameAllocator 🔒

Functions

frame_alloc

allocate a frame

frame_allocator_test

a simple test for frame allocator

frame_dealloc 🔒

deallocate a frame

init_frame_allocator

initiate the frame allocator using ekernel and MEMORY_END

Type Aliases

FrameAllocatorImpl 🔒