Module os::syscall

Implementation of syscalls

The single entry point to all system calls, syscall(), is called whenever userspace wishes to perform a system call using the ecall instruction. In this case, the processor raises an ‘Environment call from U-mode’ exception, which is handled as one of the cases in crate::trap::trap_handler.

For clarity, each single syscall is implemented as its own function, named sys_ then the name of the syscall. You can find functions like this in submodules, and you should also implement syscalls this way.

Modules

fs 🔒

gui 🔒

input 🔒

net 🔒

process 🔒

sync 🔒

thread 🔒

Constants

SYSCALL_CLOSE

close syscall

SYSCALL_CONDVAR_CREATE

condvar_create syscall

SYSCALL_CONDVAR_SIGNAL

condvar_signal syscall

SYSCALL_CONDVAR_WAIT

condvar_wait syscall

SYSCALL_CONNECT 🔒

connect syscall

SYSCALL_DUP

dup syscall

SYSCALL_ENABLE_DEADLOCK_DETECT

enable deadlock detect syscall

SYSCALL_EVENT_GET 🔒

SYSCALL_EXEC

exec syscall

SYSCALL_EXIT

exit syscall

SYSCALL_FORK

fork syscall

SYSCALL_FRAMEBUFFER 🔒

SYSCALL_FRAMEBUFFER_FLUSH 🔒

SYSCALL_FSTAT

fstat syscall

SYSCALL_GETPID

getpid syscall

SYSCALL_GETTID

gettid syscall

SYSCALL_GETTIMEOFDAY

gettimeofday syscall

SYSCALL_KEY_PRESSED 🔒

SYSCALL_KILL

kill syscall

SYSCALL_LINKAT

linkat syscall

SYSCALL_MMAP

mmap syscall

SYSCALL_MUNMAP

munmap syscall

SYSCALL_MUTEX_CREATE

mutex_create syscall

SYSCALL_MUTEX_LOCK

mutex_lock syscall

SYSCALL_MUTEX_UNLOCK

mutex_unlock syscall

SYSCALL_OPENAT

openat syscall

SYSCALL_PIPE

pipe syscall

SYSCALL_READ

read syscall

SYSCALL_SEMAPHORE_CREATE

semaphore_create syscall

SYSCALL_SEMAPHORE_DOWN

semaphore_down syscall

SYSCALL_SEMAPHORE_UP

semaphore_up syscall

SYSCALL_SET_PRIORITY

set priority syscall

SYSCALL_SLEEP

sleep syscall

SYSCALL_SPAWN

spawn syscall

SYSCALL_TASK_INFO

task info syscall

SYSCALL_THREAD_CREATE

thread_create syscall

SYSCALL_UNLINKAT

unlinkat syscall

SYSCALL_WAITPID

waitpid syscall

SYSCALL_WAITTID

waittid syscall

SYSCALL_WRITE

write syscall

SYSCALL_YIELD

yield syscall

Functions

syscall

handle syscall exception with syscall_id and other arguments