os::syscall

Module process

Functions

sys_exec

sys_exit

sys_fork

sys_get_time

sys_getpid

sys_kill

sys_waitpid

If there is not a child process whose pid is same as given, return -1. Else if there is a child process but it is still running, return -2.

sys_yield