Module os::task

Implementation of process management mechanism

Here is the entry for process scheduling required by other modules (such as syscall or clock interrupt). By suspending or exiting the current process, you can modify the process state, manage the process queue through TASK_MANAGER, and switch the control flow through PROCESSOR.

Be careful when you see __switch. Control flow around this function might not be what you expect.

Modules

action 🔒

context 🔒

Implementation of TaskContext

id 🔒

Task pid implementation.

manager 🔒

Implementation of TaskManager

processor 🔒

Implementation of Processor and Intersection of control flow

signal 🔒

switch 🔒

Wrap switch.S as a function

task 🔒

Types related to task management & Functions for completely changing TCB

Structs

INITPROC

Creation of initial process

KernelStack

Kernel stack for a process(task)

PidHandle

Abstract structure of PID

SignalAction

Action for a signal

SignalActions

Signal actions

SignalFlags

Signal flags

TaskContext

task context structure containing some registers

TaskControlBlock

Task control block structure

Enums

TaskStatus

task status: UnInit, Ready, Running, Exited

Constants

IDLE_PID

pid of usertests app in make run TEST=1

MAX_SIG

The max signal number

Functions

add_initproc

Add init process to the manager

add_task

Add process to ready queue

call_kernel_signal_handler 🔒

call kernel signal handler

call_user_signal_handler 🔒

call user signal handler

check_pending_signals 🔒

Check if the current task has any signal to handle

check_signals_error_of_current

Check if the current task has any signal to handle

current_add_signal

Add signal to the current task

current_task

Get a copy of the current task

current_trap_cx

Get the mutable reference to trap context of current task

current_user_token

Get the current user token(addr of page table)

exit_current_and_run_next

Exit the current ‘Running’ task and run the next task in task list.

handle_signals

Handle signals for the current process

kstack_alloc

allocate a new kernel stack

pid2task

Get process by pid

pid_alloc

Allocate a new PID

run_tasks

The main part of process execution and scheduling Loop fetch_task to get the process that needs to run, and switch the process through __switch

schedule

Return to idle control flow for new scheduling

suspend_current_and_run_next

Make current task suspended and switch to the next task

take_current_task

Get current task through take, leaving a None in its place